+++ man-roxygen/param-tz.R
@@ -1,0 +1,3 @@
+#' @param tz A timezone to convert dates and times to.
+#'   Read `vignette("setup", package = "ruODK")` to learn how `ruODK`'s
+#'   timezone can be set globally or per function.